+++ examples/react-app/src/components/transfer.tsx
@@ -46,25 +46,42 @@ export default function Transfer({ isSigning, setIsSigning }: Props) {
         },
       );
 
-      const result = await resp?.waitForResult();
-
       setToast({
         open: true,
         type: 'success',
-        children: (
-          <p>
-            Transfer successful! View it on the{' '}
-            <a
-              href={`https://app.fuel.network/tx/${result?.id}`}
-              className="underline"
-              target="_blank"
-              rel="noreferrer"
-            >
-              block explorer
-            </a>
-          </p>
-        ),
+        children: 'Transaction submitted!',
       });
+
+      // after 3 seconds we'll check if transaction is done
+      const TIME_TO_WAIT = 3000;
+      setTimeout(() => {
+        // execute this inside timeout to avoid block the user flow
+        async function checkResult() {
+          if (wallet) {
+            const result = await resp?.waitForResult();
+
+            setToast({
+              open: true,
+              type: 'success',
+              children: (
+                <p>
+                  Transferred successfully! View it on the{' '}
+                  <a
+                    href={`https://app.fuel.network/tx/${result.id}`}
+                    className="underline"
+                    target="_blank"
+                    rel="noreferrer"
+                  >
+                    block explorer
+                  </a>
+                </p>
+              ),
+            });
+          }
+        }
+
+        checkResult();
+      }, TIME_TO_WAIT);
     } catch (err) {
       const error = err as CustomError;
       console.error(error.message);