+++ examples/react-next/src/components/Providers.tsx
@@ -3,32 +3,19 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 
-import {
-  BakoSafeConnector,
-  FuelWalletConnector,
-  FuelWalletDevelopmentConnector,
-  FueletWalletConnector,
-} from '@fuels/connectors';
-import { WalletConnectConnector } from '@fuels/connectors/walletconnect';
+import { defaultConnectors } from '@fuels/connectors';
 import { FuelProvider } from '@fuels/react';
 
 const queryClient = new QueryClient();
 
+const fuelConfig = {
+  connectors: defaultConnectors({ devMode: true }),
+};
+
 export const Providers = ({ children }: { children: React.ReactNode }) => {
   return (
     <QueryClientProvider client={queryClient}>
-      <FuelProvider
-        theme="dark"
-        fuelConfig={{
-          connectors: [
-            new FuelWalletConnector(),
-            new FuelWalletDevelopmentConnector(),
-            new BakoSafeConnector(),
-            new FueletWalletConnector(),
-            new WalletConnectConnector(),
-          ],
-        }}
-      >
+      <FuelProvider theme="dark" fuelConfig={fuelConfig}>
         {children}
       </FuelProvider>