+++ packages/connectors/src/defaultConnectors.ts
@@ -1,24 +1,44 @@
 import { BakoSafeConnector } from '@fuel-connectors/bako-safe';
-import { BurnerWalletConnector } from '@fuel-connectors/burner-wallet-connector';
+import {
+  type BurnerWalletConfig,
+  BurnerWalletConnector,
+} from '@fuel-connectors/burner-wallet-connector';
 import { FuelWalletDevelopmentConnector } from '@fuel-connectors/fuel-development-wallet';
 import { FuelWalletConnector } from '@fuel-connectors/fuel-wallet';
 import { FueletWalletConnector } from '@fuel-connectors/fuelet-wallet';
+import { SolanaConnector } from '@fuel-connectors/solana-connector';
 import { WalletConnectConnector } from '@fuel-connectors/walletconnect-connector';
+import type { Config } from '@wagmi/core';
+import type { ProviderType } from '@web3modal/solana/dist/types/src/utils/scaffold';
 import type { FuelConnector } from 'fuels';
 
 type DefaultConnectors = {
   devMode?: boolean;
+  wcProjectId?: string;
+  burnerWalletConfig?: BurnerWalletConfig;
+  ethWagmiConfig?: Config;
+  solanaConfig?: ProviderType;
 };
 
 export function defaultConnectors({
   devMode,
+  wcProjectId,
+  burnerWalletConfig,
+  ethWagmiConfig,
+  solanaConfig: _solanaConfig,
 }: DefaultConnectors = {}): Array<FuelConnector> {
-  const connectors = [
+  const connectors: Array<FuelConnector> = [
     new FuelWalletConnector(),
     new BakoSafeConnector(),
     new FueletWalletConnector(),
-    new WalletConnectConnector(),
-    new BurnerWalletConnector(),
+    new WalletConnectConnector({
+      projectId: wcProjectId,
+      wagmiConfig: ethWagmiConfig,
+    }),
+    new SolanaConnector({
+      projectId: wcProjectId,
+    }),
+    new BurnerWalletConnector(burnerWalletConfig),
   ];
 
   if (devMode) {