+++ packages/connectors/src/index.ts
@@ -5,3 +5,4 @@ export * from '@fuel-connectors/fuelet-wallet';
 export * from '@fuel-connectors/bako-safe';
 export * from '@fuel-connectors/burner-wallet-connector';
 export * from '@fuel-connectors/walletconnect-connector';
+export * from '@fuel-connectors/solana-connector';