+++ packages/evm-connector/src/test/evmWalletConnector.test.ts
@@ -111,6 +111,8 @@ describe('EVM Wallet Connector', () => {
       launchNodeOptions: {
         args: ['--snapshot', snapshotPath],
         loggingEnabled: false,
+        // use fixed port to don't conflict with other packages,
+        port: '4001',
       },
     });
 
@@ -327,6 +329,8 @@ describe('EVM Wallet Connector', () => {
 
       await connector.sendTransaction(accountAddress, transactionRequest);
 
+      await new Promise((resolve) => setTimeout(resolve, 100));
+
       // Check balances are correct
       const predicateAltBalanceFinal =
         await createdPredicate.getBalance(ALT_ASSET_ID);
@@ -407,6 +411,8 @@ describe('EVM Wallet Connector', () => {
       await connector.accounts();
 
       await connector.sendTransaction(accountAddress, transactionRequest);
+
+      await new Promise((resolve) => setTimeout(resolve, 100));
 
       // Check balances are correct
       const predicateAltBalanceFinal =