+++ packages/solana-connector/src/test/solanaConnector.test.ts
@@ -1,0 +1,111 @@
+import path from 'node:path';
+import { launchNodeAndGetWallets } from '@fuel-ts/account/test-utils';
+import type { Asset, Network } from 'fuels';
+import { beforeAll, describe, expect, test } from 'vitest';
+import { SolanaConnector } from '../SolanaConnector';
+
+describe('Solana Connector', () => {
+  const projectId = '0000';
+
+  const connector = new SolanaConnector({ projectId });
+
+  const snapshotPath = path.join(__dirname, '');
+
+  beforeAll(async () => {
+    process.env.GENESIS_SECRET =
+      '0x6e48a022f9d4ae187bca4e2645abd62198ae294ee484766edbdaadf78160dc68';
+    await launchNodeAndGetWallets({
+      launchNodeOptions: {
+        args: ['--snapshot', snapshotPath],
+        loggingEnabled: false,
+        // use fixed port to don't conflict with other packages,
+        port: '4002',
+      },
+    });
+  });
+
+  describe('currentAccount()', () => {
+    test('throws error', async () => {
+      await expect(() => connector.currentAccount()).rejects.toThrowError(
+        'No connected accounts',
+      );
+    });
+  });
+
+  describe('signMessage()', () => {
+    test('throws error', async () => {
+      await expect(() =>
+        connector.signMessage('address', 'message'),
+      ).rejects.toThrowError('A predicate account cannot sign messages');
+    });
+  });
+
+  describe('assets()', () => {
+    test('returns an empty array', async () => {
+      expect(await connector.assets()).to.deep.equal([]);
+    });
+  });
+
+  describe('addAsset()', () => {
+    test('throws error', async () => {
+      const asset: Asset = {
+        name: '',
+        symbol: '',
+        icon: '',
+        networks: [],
+      };
+      await expect(() => connector.addAsset(asset)).rejects.toThrowError(
+        'Method not implemented.',
+      );
+    });
+  });
+
+  describe('addAssets()', () => {
+    test('throws error', async () => {
+      await expect(() => connector.addAssets([])).rejects.toThrowError(
+        'Method not implemented.',
+      );
+    });
+  });
+
+  describe('addAbi()', () => {
+    test('throws error', async () => {
+      await expect(() => connector.addAbi({})).rejects.toThrowError(
+        'Method not implemented.',
+      );
+    });
+  });
+
+  describe('getAbi()', () => {
+    test('throws error', async () => {
+      await expect(() => connector.getAbi('contractId')).rejects.toThrowError(
+        'Cannot get contractId ABI for a predicate',
+      );
+    });
+  });
+
+  describe('hasAbi()', () => {
+    test('throws error', async () => {
+      await expect(() => connector.hasAbi('contractId')).rejects.toThrowError(
+        'A predicate account cannot have an ABI',
+      );
+    });
+  });
+
+  describe('addNetwork()', () => {
+    test('throws error', async () => {
+      await expect(() => connector.addNetwork('')).rejects.toThrowError(
+        'Method not implemented.',
+      );
+    });
+  });
+
+  describe('selectNetwork()', () => {
+    test('throws error', async () => {
+      const network: Network = { url: '', chainId: 0 };
+      await expect(() => connector.selectNetwork(network)).rejects.toThrowError(
+        'Method not implemented.',
+      );
+    });
+  });
+});